youtube_crew/pyproject.toml
[project]
name = "youtube_crew"
version = "0.1.0"
description = "youtube_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.3",
    "edge-tts>=7.2.8",
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "playwright>=1.58.0",
    "selenium>=4.43.0",
]

[project.scripts]
youtube_crew = "youtube_crew.main:run"
run_crew = "youtube_crew.main:run"
train = "youtube_crew.main:train"
replay = "youtube_crew.main:replay"
test = "youtube_crew.main:test"
run_with_trigger = "youtube_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
